Mega-pass checkpoint: integration round 1 green
Wires the ~88 jampgame fn-skeleton files landed pre-integration into a
compiling workspace. Mechanical fixes: prelude re-exports for consts/enums
already ported elsewhere (WP_*/FP_*/TEAM_*/CHAN_*/entity_type/stat_index/
means_of_death/pd_sounds/force_hand_anims/pers_enum/pmtype/broken_limb/
task_id_t/trackchan/wl_e/bSet_t/class_t/Eorientations/ammo_t glob imports),
per-file `use crate::<module>::{Fn}` wiring for same-crate helpers, qtrue/
qfalse local consts per established convention, a handful of small missing
#define-derived consts transcribed straight from the oracle headers
(FORCE_LEVEL_*, ENTITYNUM_*, MOVER_*, SETANIM_*, entity-effect EF_*/EF2_*
flags, itemType_t), two ABI struct field-name fixes to match Raven exactly
(trajectory_t trBase/trDelta/trDuration/trType/trTime, usercmd_t serverTime),
two stale `*mut c_void` vtable-adjacent fields resolved now that their
pointee types landed (bgEntity_t::m_pVehicle -> Vehicle_t, Vehicle_t::
m_pVehicleInfo -> vehicleInfo_t), replacement of bogus `extern "C" { .. }`
blocks (treated intra-crate Rust fns as link-time externs) with ordinary
`use`, and Sync/ZeroValid/Copy marker additions for a few pre-existing gaps
blocking the build (BG_field_t, u8, vehicleType_t, navInfo_t).

Found and fixed a correctness bug, not just a compile gap: many
`animNumber_t` (`BOTH_*`/`TORSO_*`/`LEGS_*`) match arms silently
type-checked as irrefutable catch-all bindings (animNumber_t is a
documented-deferred ~1500-entry enum), making those predicates
unconditionally return their first arm's value. Parked instead of invented
per porting-rules SA2 — animNumber_t is out of scope for this round.

Everything else needing real design judgment is parked (`todo!()` +
`PORT-ESCALATION(topic)`) rather than guessed: ambient "current NPC"/
NPCInfo/level/ucmd globals with no GameContext channel yet (`ai-context`),
functions needing a GameContext receiver their faithful C signature carries
none of (`seam-threading`), and reads of genuinely unported runtime data
tables (`bg_itemlist`/`ammoData`/`weaponData`/`forcePowerNeeded`/
`bgSiegeClasses`/`g_vehicleInfo`/`bgToggleableSurfaces`) that are a real
fork-discovery-ruling-1 placement question, not a missing `use`.

cargo build --workspace is green.

+++ crates/mp/game/src/NPC_AI_Grenadier.rs
@@ -9,6 +9,11 @@
 #![allow(non_snake_case, unused, clippy::all)]
 
 use crate::prelude::*;
+use crate::NPC_combat::G_ClearEnemy;
+use crate::NPC_reactions::NPC_Pain;
+use crate::NPC_sounds::G_AddVoiceEvent;
+use crate::g_timer::TIMER_Set;
+use crate::q_math::Q_irand;
 use mp_bg::public::entity_event::entity_event_t::{EV_CONFUSE1, EV_CONFUSE3, EV_PUSHED1, EV_PUSHED3};
 use mp_qshared::common::mp::qcommon::b_state_t::bState_t;
 
@@ -68,7 +73,7 @@ pub fn NPC_Grenadier_PlayConfusionSound(self_: *mut gentity_t) {
         TIMER_Set(self_, c"flee".as_ptr() as *const c_char, 0);
         let npc = (*self_).NPC as *mut gNPC_t;
         (*npc).squadState = SQUAD_IDLE;
-        (*npc).tempBehavior = bState_t::BS_DEFAULT as i32;
+        (*npc).tempBehavior = bState_t::BS_DEFAULT;
         G_ClearEnemy(self_); // FIXME: or just self->enemy = NULL;? (Raven comment).
         (*npc).investigateCount = 0;
     }
